pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]

[project]
name = "hlm12erc"
version = "0.0.1"
authors = [{ name = "Hudson Mendes", email = "hlm12@student.london.ac.uk" }]
description = "Final Project from University of London"
readme = "README.md"
license = { file = "LICENSE" }
urls = { homepage = "https://github.com/hudsonmendes/cm3070-fp" }
keywords = ["university-of-london"]
dependencies = [
    "torch >= 2.0.1",
    "torchtext >= 0.15.2",
    "torchvision >= 0.15.2",
    "transformers >= 4.30.2",
    "Pillow >= 10.0.0",
    "scikit-learn >= 1.3.0",
    "opencv-python>=4.8.0.76",
]

[project.optional-dependencies]
dev = ["pre-commit>=3.3.3", "black[jupyter]>=23.7.0", "isort>=5.12.0"]
test = ["pytest>=7.4.0", "hypothesis>=6.84.2"]
etl = [
    "kaggle>=1.5.13",
    "tqdm>=4.65.0",
    "pandas>=2.0.1",
    "google-cloud-storage>=2.10.0",
    "moviepy>=1.0.3",
]
eda = [
    "gensim",
    "tensorflow",
    "tensorflow-hub",
    "torch",
    "transformers",
    "librosa",
    "umap-learn",
    "matplotlib",
    "wordcloud",
    "pyLDAvis",
    "nltk",
]
modelling = []
training = [
    "pandas>=2.0.1",
    "PyYAML>=6.0",
    "transformers[torch]>=4.30.2",
    "wandb>=0.15.8",
]
serving = ["fire>=0.5.0"]

[tool.black]
line-length = 120

[tool.isort]
profile = "black"
line_length = 120
src_paths = ["dev", "src", "tests"]
known_third_party = [
    "numpy",
    "pandas",
    "tqdm",
    "pytest",
    "torch",
    "kaggle",
    "moviepy",
    "PIL",
]
known_first_party = ["hlm12erc"]
import_heading_stdlib = "Python Built-in Modules"
import_heading_thirdparty = "Third-Party Libraries"
import_heading_firstparty = "My Packages and Modules"
import_heading_localfolder = "Local Folders"
